Change CustomPropertyID to uint32_t. Fix SDL example.
64-bit integer not supported in JS

+++ include/cobbletext/common.hpp
@@ -7,6 +7,6 @@ namespace cobbletext {
 typedef uint32_t FontID;
 typedef uint32_t GlyphID;
 typedef uint32_t InlineObjectID;
-typedef uint64_t CustomPropertyID;
+typedef uint32_t CustomPropertyID;
 
 }
